requirements.txt
selenium
pandas
linkedin-api
python-dotenv
webdriver_manager
beautifulsoup4
requests